all modules compile again

+++ persistence/src/main/java/org/smallmind/persistence/cache/ByReferenceDistributedCacheDao.java
@@ -1,27 +1,36 @@
 package org.smallmind.persistence.cache;
 
-import java.util.LinkedList;
+import java.util.Comparator;
 import org.smallmind.persistence.Durable;
-import org.smallmind.persistence.DurableVector;
+import org.smallmind.persistence.DurableKey;
 import org.smallmind.persistence.VectorKey;
 import org.smallmind.persistence.VectoredDao;
+import org.smallmind.persistence.cache.util.CachedList;
 
-public class ByReferenceDistributedCacheDao<Long, D extends Durable<Long>> extends WaterfallCacheDao<Long, D> {
+public class ByReferenceDistributedCacheDao<I extends Comparable<I>, D extends Durable<I>> extends WaterfallCacheDao<I, D> {
 
-   public ByReferenceDistributedCacheDao (CacheDomain<Long, D> cacheDomain) {
+   public ByReferenceDistributedCacheDao (CacheDomain<I, D> cacheDomain) {
 
       super(null, cacheDomain);
    }
 
-   public ByReferenceDistributedCacheDao (VectoredDao<Long, D> nextDao, CacheDomain<Long, D> cacheDomain) {
+   public ByReferenceDistributedCacheDao (VectoredDao<I, D> nextDao, CacheDomain<I, D> cacheDomain) {
 
       super(nextDao, cacheDomain);
    }
 
-   public D get (Class<D> durableClass, Long id) {
+   public D acquire (Class<D> durableClass, I id) {
+
+      DurableKey<I, D> durableKey = new DurableKey<I, D>(durableClass, id);
+
+      return getInstanceCache(durableClass).get(durableKey.getKey());
+   }
+
+   public D get (Class<D> durableClass, I id) {
 
       D cachedDurable;
-      VectoredDao<Long, D> nextDao = getNextDao();
+      VectoredDao<I, D> nextDao = getNextDao();
+      DurableKey<I, D> durableKey = new DurableKey<I, D>(durableClass, id);
 
       if (nextDao != null) {
          if ((cachedDurable = nextDao.get(durableClass, id)) != null) {
@@ -30,7 +39,7 @@ public class ByReferenceDistributedCacheDao<Long, D extends Durable<Long>> exten
          }
       }
 
-      if ((cachedDurable = getInstanceCache(durableClass).get(id)) != null) {
+      if ((cachedDurable = getInstanceCache(durableClass).get(durableKey.getKey())) != null) {
          if (nextDao != null) {
             nextDao.persist(durableClass, cachedDurable);
          }
@@ -43,15 +52,20 @@ public class ByReferenceDistributedCacheDao<Long, D extends Durable<Long>> exten
 
    public D persist (Class<D> durableClass, D durable) {
 
-      D cachedDurable;
-      VectoredDao<Long, D> nextDao = getNextDao();
+      if (durable != null) {
 
-      if ((cachedDurable = getInstanceCache(durableClass).putIfAbsent(durable.getId(), durable)) != null) {
-         return cachedDurable;
-      }
+         D cachedDurable;
+         VectoredDao<I, D> nextDao = getNextDao();
+         DurableKey<I, D> durableKey = new DurableKey<I, D>(durableClass, durable.getId());
 
-      if (nextDao != null) {
-         nextDao.persist(durableClass, durable);
+         if ((cachedDurable = getInstanceCache(durableClass).putIfAbsent(durableKey.getKey(), durable)) != null) {
+
+            return cachedDurable;
+         }
+
+         if (nextDao != null) {
+            nextDao.persist(durableClass, durable);
+         }
       }
 
       return durable;
@@ -59,47 +73,57 @@ public class ByReferenceDistributedCacheDao<Long, D extends Durable<Long>> exten
 
    public void delete (Class<D> durableClass, D durable) {
 
-      VectoredDao<Long, D> nextDao = getNextDao();
+      if (durable != null) {
 
-      getInstanceCache(durableClass).remove(durable.getId());
+         VectoredDao<I, D> nextDao = getNextDao();
+         DurableKey<I, D> durableKey = new DurableKey<I, D>(durableClass, durable.getId());
 
-      if (nextDao != null) {
-         nextDao.delete(durableClass, durable);
+         getInstanceCache(durableClass).remove(durableKey.getKey());
+
+         if (nextDao != null) {
+            nextDao.delete(durableClass, durable);
+         }
       }
    }
 
-   public void updateInVector (VectorKey<Long, D> vectorKey, D durable) {
+   public void updateInVector (VectorKey<D> vectorKey, D durable) {
 
-      DurableVector<Long, D> vector;
-      VectoredDao<Long, D> nextDao = getNextDao();
+      if (durable != null) {
 
-      if ((vector = getVectorCache(vectorKey.getElementClass()).get(vectorKey.getKey())) != null) {
-         vector.add(durable);
-      }
+         DurableVector<I, D> vector;
+         VectoredDao<I, D> nextDao = getNextDao();
 
-      if (nextDao != null) {
-         nextDao.updateInVector(vectorKey, durable);
-      }
-   }
+         if ((vector = getVectorCache(vectorKey.getElementClass()).get(vectorKey.getKey())) != null) {
+            vector.add(durable);
+         }
 
-   public void removeFromVector (VectorKey<Long, D> vectorKey, D durable) {
-
-      DurableVector<Long, D> vector;
-      VectoredDao<Long, D> nextDao = getNextDao();
-
-      if ((vector = getVectorCache(vectorKey.getElementClass()).get(vectorKey.getKey())) != null) {
-         vector.remove(durable);
-      }
-
-      if (nextDao != null) {
-         nextDao.removeFromVector(vectorKey, durable);
+         if (nextDao != null) {
+            nextDao.updateInVector(vectorKey, durable);
+         }
       }
    }
 
-   public DurableVector<Long, D> getVector (VectorKey<Long, D> vectorKey) {
+   public void removeFromVector (VectorKey<D> vectorKey, D durable) {
 
-      DurableVector<Long, D> cachedVector;
-      VectoredDao<Long, D> nextDao = getNextDao();
+      if (durable != null) {
+
+         DurableVector<I, D> vector;
+         VectoredDao<I, D> nextDao = getNextDao();
+
+         if ((vector = getVectorCache(vectorKey.getElementClass()).get(vectorKey.getKey())) != null) {
+            vector.remove(durable);
+         }
+
+         if (nextDao != null) {
+            nextDao.removeFromVector(vectorKey, durable);
+         }
+      }
+   }
+
+   public DurableVector<I, D> getVector (VectorKey<D> vectorKey) {
+
+      DurableVector<I, D> cachedVector;
+      VectoredDao<I, D> nextDao = getNextDao();
 
       if (nextDao != null) {
          if ((cachedVector = nextDao.getVector(vectorKey)) != null) {
@@ -119,40 +143,85 @@ public class ByReferenceDistributedCacheDao<Long, D extends Durable<Long>> exten
       return null;
    }
 
-   public DurableVector<Long, D> persistVector (VectorKey<Long, D> vectorKey, Iterable<D> elementIter) {
+   public DurableVector<I, D> persistVector (VectorKey<D> vectorKey, DurableVector<I, D> vector) {
 
-      DurableVector<Long, D> vector;
-      DurableVector<Long, D> cachedVector;
+      DurableVector<I, D> migratedVector;
+      DurableVector<I, D> cachedVector;
+      VectoredDao<I, D> nextDao = getNextDao();
 
-      LinkedList<D> cacheConsistentElements;
-      D inCacheDurable;
-      VectoredDao<Long, D> nextDao = getNextDao();
+      migratedVector = migrateVector(vector);
+      if ((cachedVector = getVectorCache(vectorKey.getElementClass()).putIfAbsent(vectorKey.getKey(), migratedVector)) != null) {
 
-      cacheConsistentElements = new LinkedList<D>();
-      for (D element : elementIter) {
-         if ((inCacheDurable = getInstanceCache(vectorKey.getElementClass()).putIfAbsent(element.getId(), element)) != null) {
-            cacheConsistentElements.add(inCacheDurable);
-         }
-         else {
-            cacheConsistentElements.add(element);
-         }
-      }
-
-      if ((cachedVector = getVectorCache(vectorKey.getElementClass()).putIfAbsent(vectorKey.getKey(), vector = new ByReferenceDurableVector<Long, D>(cacheConsistentElements))) != null) {
          return cachedVector;
       }
-      else {
-         if (nextDao != null) {
-            nextDao.persistVector(vectorKey, cacheConsistentElements);
-         }
+      else if (nextDao != null) {
+         nextDao.persistVector(vectorKey, migratedVector);
       }
 
       return vector;
    }
 
-   public void deleteVector (VectorKey<Long, D> vectorKey) {
+   public DurableVector<I, D> migrateVector (DurableVector<I, D> vector) {
 
-      VectoredDao<Long, D> nextDao = getNextDao();
+      if (vector.isSingular()) {
+         if (!(vector instanceof SingularByReferenceDurableVector)) {
+
+            return new SingularByReferenceDurableVector<I, D>(vector.head(), vector.getTimeToLive());
+         }
+
+         return vector;
+      }
+      else {
+
+         if (!(vector instanceof ByReferenceDurableVector)) {
+
+            return new ByReferenceDurableVector<I, D>(new CachedList<D>(vector.asList()), vector.getComparator(), vector.getMaxSize(), vector.getTimeToLive(), vector.isOrdered());
+         }
+
+         return vector;
+      }
+   }
+
+   public DurableVector<I, D> createSingularVector (VectorKey<D> vectorKey, D durable, long timeToLive) {
+
+      DurableKey<I, D> durableKey;
+      D inCacheDurable;
+
+      durableKey = new DurableKey<I, D>(vectorKey.getElementClass(), durable.getId());
+      if ((inCacheDurable = getInstanceCache(vectorKey.getElementClass()).putIfAbsent(durableKey.getKey(), durable)) != null) {
+
+         return new SingularByReferenceDurableVector<I, D>(inCacheDurable, timeToLive);
+      }
+
+      return new SingularByReferenceDurableVector<I, D>(durable, timeToLive);
+   }
+
+   public DurableVector<I, D> createVector (VectorKey<D> vectorKey, Iterable<D> elementIter, Comparator<D> comparator, int maxSize, long timeToLive, boolean ordered) {
+
+      CachedList<D> cacheConsistentElements;
+      DurableKey<I, D> durableKey;
+      D inCacheDurable;
+
+      cacheConsistentElements = new CachedList<D>();
+      for (D element : elementIter) {
+         if (element != null) {
+
+            durableKey = new DurableKey<I, D>(vectorKey.getElementClass(), element.getId());
+            if ((inCacheDurable = getInstanceCache(vectorKey.getElementClass()).putIfAbsent(durableKey.getKey(), element)) != null) {
+               cacheConsistentElements.add(inCacheDurable);
+            }
+            else {
+               cacheConsistentElements.add(element);
+            }
+         }
+      }
+
+      return new ByReferenceDurableVector<I, D>(cacheConsistentElements, comparator, maxSize, timeToLive, ordered);
+   }
+
+   public void deleteVector (VectorKey<D> vectorKey) {
+
+      VectoredDao<I, D> nextDao = getNextDao();
 
       getVectorCache(vectorKey.getElementClass()).remove(vectorKey.getKey());
